Refactored the File-Parsing
Refactored the File-Parsing, so it happens in the ParseControlle now.
Added a DpsReportControlle for uploading to DPSReport.
Added a RelayCommand for the few times when a command is needed.
